Splitter opp K9OppslagGateway i ArbeidsgivereGateway, BarnGateway og SokerGateway; Sletter spor av sparkel
