+++ src/main/kotlin/no/nav/helse/general/systemauth/AccessTokenClientResolver.kt
@@ -44,7 +44,6 @@ internal class AccessTokenClientResolver(
     )
 
     internal fun aktoerRegister() = naisStsAccessTokenClient
-    internal fun sparkel() = naisStsAccessTokenClient
     internal fun k9Oppslag() = naisStsAccessTokenClient
     internal fun pleiepengesoknadMottak() = azureV2AccessTokenClient
 }
