+++ src/main/kotlin/no/nav/helse/soker/SokerApis.kt
@@ -8,7 +8,6 @@ import io.ktor.response.respond
 import io.ktor.routing.Route
 import no.nav.helse.general.auth.getNorskIdent
 import no.nav.helse.general.getCallId
-import no.nav.helse.k9.K9OppslagSokerService
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 
@@ -17,14 +16,14 @@ private val logger: Logger = LoggerFactory.getLogger("nav.sokerApis")
 
 @KtorExperimentalLocationsAPI
 fun Route.sokerApis(
-    k9OppslagSokerService: K9OppslagSokerService
+    sokerService: SokerService
 ) {
 
     @Location("/soker")
     class getSoker
 
     get { _: getSoker ->
-        call.respond(k9OppslagSokerService.getSoker(
+        call.respond(sokerService.getSoker(
             ident = call.getNorskIdent().getValue(),
             callId = call.getCallId()
         ))
