+++ src/main/resources/application.conf
@@ -15,8 +15,6 @@ nav {
         addresses = ${?CORS_ADDRESSES}
     }
     gateways {
-        sparkel_url = ""
-        sparkel_url = ${?SPARKEL_URL}
         aktoer_register_url =""
         aktoer_register_url = ${?AKTOER_REGISTER_URL},
         pleiepengesoknad_mottak_base_url = "",
