+++ src/test/kotlin/no/nav/helse/ApplicationTest.kt
@@ -40,8 +40,9 @@ class ApplicationTest {
             .pleiepengesoknadApiConfig()
             .build()
             .stubPleiepengerDokumentHealth()
-            .stubSparkelIsReady()
             .stubPleiepengesoknadMottakHealth()
+            .stubOppslagHealth()
+            .stubAktoerRegisterHealth()
             .stubAktoerRegisterGetAktoerId()
             .stubLeggSoknadTilProsessering()
             .stubPleiepengerDokument()
