+++ src/test/kotlin/no/nav/helse/ApplicationWithMocks.kt
@@ -27,7 +27,6 @@ class ApplicationWithMocks {
                 .pleiepengesoknadApiConfig()
                 .build()
                 .stubPleiepengerDokumentHealth()
-                .stubSparkelIsReady()
                 .stubPleiepengesoknadMottakHealth()
                 .stubAktoerRegisterGetAktoerId()
                 .stubLeggSoknadTilProsessering()
